api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
pdfplumber>=0.11.0
google-generativeai>=0.4.0
jinja2>=3.1.0
python-multipart>=0.0.9
supabase>=2.3.0
Pillow>=10.2.0
pillow-heif>=0.15.0
PyMuPDF>=1.23.0
SQLAlchemy>=2.0.0
